docs: 불필요한 코드 제거 및 버전 업 환경세팅
