+++ lib/common/const/data.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 const ACCESS_TOKEN_KEY = 'ACCESS_TOKEN';
@@ -7,14 +5,11 @@ const REFRESH_TOKEN_KEY = 'REFRESH_TOKEN';
 
 const emailPattern = r'^[a-zA-Z0-9+-\_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$';
 const passwordPattern =
-    // r'^(?=.*[A-Za-z])(?=.*\d)(?=.*[@!%*#?&])[A-Za-z\d@!%*#?&]{8,}$';
     r"^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d@!%*#?&:]{8,}$";
 
 // localhost
 final emulatorIp = '10.0.2.2:8000/api/v1';
 final simulatorIp = '127.0.0.1:8000/api/v1';
-
-// final ip = Platform.isIOS ? simulatorIp : emulatorIp;
 
 // baseUrl
 final baseUrl = 'https://mohim-api.com/api/v1';
