+++ lib/group/model/gathering.dart
@@ -2,7 +2,6 @@ import 'package:json_annotation/json_annotation.dart';
 
 part 'gathering.g.dart';
 
-//flutter pub run build_runner build
 @JsonSerializable()
 class Gathering {
   int id;
