+++ lib/group/view/group_screen.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:phonebook/common/const/colors.dart';
 import 'package:phonebook/common/const/style.dart';
-import 'package:phonebook/group/controller/group_controller.dart';
 import 'package:phonebook/group/view/gathering_list_screen.dart';
 import 'package:phonebook/group/view/ministry_list_screen.dart';
 import 'package:phonebook/group/view/parish_list_screen.dart';
@@ -76,8 +74,6 @@ class _GroupScreenState extends State<GroupScreen>
                 width: 0.0,
               ),
             ),
-            // borderRadius: BorderRadius.only(
-            //     topLeft: Radius.circular(8.0), topRight: Radius.circular(8.0)),
           ),
           child: TabBar(
             controller: _tabController,
