+++ lib/main.dart
@@ -20,7 +20,6 @@ void main() async {
 class _App extends StatelessWidget {
   const _App({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
