+++ lib/profile/view/image_screen.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:phonebook/common/const/colors.dart';
 import 'package:phonebook/common/layout/default_layout.dart';
-import 'package:phonebook/contact/controller/member_controller.dart';
 import 'package:phonebook/profile/controller/profile_controller.dart';
 import 'package:phonebook/profile/util/cache_manager.dart';
 
@@ -25,11 +24,7 @@ class ImageScreen extends StatelessWidget {
           child: Center(
             child: Obx(() {
               return Get.find<ProfileController>().imageUrl.isNotEmpty
-                  ?
-                  // InteractiveViewer(
-                  //     child: Image.network(
-                  //         Get.find<ProfileController>().imageUrl.value))
-                  CachedNetworkImage(
+                  ? CachedNetworkImage(
                       imageUrl: Get.find<ProfileController>().imageUrl.value,
                       errorWidget: (context, url, error) => Icon(
                         Icons.person,
