+++ lib/user/component/auth_modal.dart
@@ -1,4 +1,3 @@
-import 'package:phonebook/common/component/custom_rounded_button.dart';
 import 'package:phonebook/common/const/colors.dart';
 import 'package:flutter/material.dart';
