+++ lib/user/model/login_response.dart
@@ -1,11 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 
-/// 이 구문은 `User` 클래스가 생성된 파일의 private 멤버들을
-/// 접근할 수 있도록 해줍니다. 여기에는 *.g.dart 형식이 들어갑니다.
-/// * 에는 소스 파일의 이름이 들어갑니다.
 part 'login_response.g.dart';
 
-/// 코드 생성기에 이 클래스가 JSON 직렬화 로직이 만들어져야 한다고 알려주는 어노테이션입니다.
 @JsonSerializable()
 class LoginResponse {
   LoginResponse(
@@ -15,13 +11,8 @@ class LoginResponse {
   @JsonKey(name: 'access_token')
   String accessToken;
 
-  /// map에서 새로운 User 인스턴스를 생성하기 위해 필요한 팩토리 생성자입니다.
-  /// 생성된 `_$UserFromJson()` 생성자에게 map을 전달해줍니다.
-  /// 생성자의 이름은 클래스의 이름을 따릅니다. 본 예제의 경우 User를 따릅니다.
   factory LoginResponse.fromJson(Map<String, dynamic> json) =>
       _$LoginResponseFromJson(json);
 
-  /// `toJson`은 클래스가 JSON 인코딩의 지원을 선언하는 규칙입니다.
-  /// 이의 구현은 생성된 private 헬퍼 메서드 `_$UserToJson`을 단순히 호출합니다.
   Map<String, dynamic> toJson() => _$LoginResponseToJson(this);
 }
