+++ lib/user/view/sign_up_screen.dart
@@ -94,8 +94,6 @@ class _SignUpScreenState extends State<SignUpScreen> {
                     validator: (value) {
                       final regex = RegExp(emailPattern);
                       if (value == null || value.isEmpty) {
-                        // emailValidated = false;
-                        // return '입력창이 비어있습니다.';
                       } else if (!regex.hasMatch(value)) {
                         emailValidated = false;
                         return '올바른 이메일 형식이 아닙니다.';
@@ -118,8 +116,6 @@ class _SignUpScreenState extends State<SignUpScreen> {
                     validator: (value) {
                       final regex = RegExp(passwordPattern);
                       if (value == null || value.isEmpty) {
-                        // passwordValidated = false;
-                        // return '입력창이 비어있습니다.';
                       } else if (!regex.hasMatch(value)) {
                         passwordValidated = false;
                         return '영어, 숫자를 포함하여 8자 이상으로 입력하세요.';
@@ -142,8 +138,6 @@ class _SignUpScreenState extends State<SignUpScreen> {
                     validator: (value) {
                       final regex = RegExp(passwordPattern);
                       if (value == null || value.isEmpty) {
-                        // confirmationValidated = false;
-                        // return '입력창이 비어있습니다.';
                       } else if (!regex.hasMatch(value)) {
                         confirmationValidated = false;
                         return '영어, 숫자를 포함하여 8자 이상으로 입력하세요.';
