+++ pubspec.yaml
@@ -17,7 +17,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 0.1.5+9
+version: 0.1.6+10
 
 environment:
   sdk: '>=2.18.6 <3.0.0'
@@ -87,7 +87,6 @@ dev_dependencies:
 flutter:
   uses-material-design: true
   assets:
-    - asset/img/misc/
     - asset/svg/icon/
     - asset/svg/logo/
     - asset/img/logo/
